[Fixes #66] Bug: not for all geonode object types we are able to run --filter
